fix: 라벨 규칙 lib 삭제 및 BODYTEXT 본문으로 수정 (#47)
